requirements.txt
watchdog
flask
amazon-textract-textractor
